** fixed some test errors
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@5517 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/TRANSFORMATIONS/FEATUREFINDER/FeatureFinder_impl.h
@@ -100,7 +100,7 @@ namespace OpenMS
 		{
 			//index
 			Size spectrum_index = input_map.RTBegin(features[i].getRT()) - input_map.begin();
-			features[i].setMetaValue("spectrum_index", spectrum_index);
+			features[i].setMetaValue("spectrum_index",(UInt) spectrum_index);
 			//native id
 			String native_id = input_map[spectrum_index].getNativeID();
 			features[i].setMetaValue("spectrum_native_id", native_id);
